val retour précisée en commentaire

+++ contains/contains.h
@@ -1,4 +1,6 @@
-size_t contains(int vect[],size_t size,int elem) {
-	//Ecrire le code de la fonction
+int contains(int vect[],size_t size,int elem) {
+	// Ecrire le code de la fonction contains
+	// la fonction renvoie l'indice de la première occurence de l'élément si il est présent dans le vecteur
+	// la fonction renvoie -1 si l'élément n'est pas trouvé
 	return -1;
 }
